feat : 플레이어 스탯 UI 및 파티클 구현

+++ Assets/Scripts/Enemy.cs
@@ -11,8 +11,8 @@ public class Enemy : MonoBehaviour
 
     [Header("Spec")]
     [SerializeField]
-    public int maxhp;
-    public int curhp;
+    public float maxhp;
+    public float curhp;
     [SerializeField]
     public int damage;
     [SerializeField]
